+++ kitty/options/definition.py
@@ -2518,7 +2518,9 @@ environment variables are expanded recursively, so if you use::
     env MYVAR1=a
     env MYVAR2=${MYVAR1}/${HOME}/b
 
-The value of MYVAR2 will be :code:`a/<path to home directory>/b`.
+The value of MYVAR2 will be :code:`a/<path to home directory>/b`. Using
+:code:`VAR=` will set it to the empty string and using just :code:`VAR`
+will delete the variable from the child process' environment.
 '''
     )